Overall refactoring (#18)
* Fix proj

* Change project organisation

* Services definition refactoring & documentations

* Models refactoring & documentations

* Forms application refactoring & documentations

* Fix documentation mistake
