+++ FacesToSmileys.ViewModels/packages.config
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="reactiveui" version="7.1.0" targetFramework="portable45-net45+win8+wpa81" />
+  <package id="reactiveui-core" version="7.1.0" targetFramework="portable45-net45+win8+wpa81" />
+  <package id="Rx-Core" version="2.2.5" targetFramework="portable45-net45+win8+wpa81" />
+  <package id="Rx-Interfaces" version="2.2.5" targetFramework="portable45-net45+win8+wpa81" />
+  <package id="Rx-Linq" version="2.2.5" targetFramework="portable45-net45+win8+wpa81" />
+  <package id="Rx-Main" version="2.2.5" targetFramework="portable45-net45+win8+wpa81" />
+  <package id="Rx-PlatformServices" version="2.2.5" targetFramework="portable45-net45+win8+wpa81" />
+  <package id="Splat" version="1.4.0" targetFramework="portable45-net45+win8+wpa81" />
+</packages>
