--- FacesToSmileys/App.xaml
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Application xmlns="http://xamarin.com/schemas/2014/forms" xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" x:Class="FacesToSmileys.App">
-    <Application.Resources>
-        <!-- Application resource dictionary -->
-    </Application.Resources>
-</Application>
