+++ FacesToSmileys/Dependencies/ExternalModule.cs
@@ -4,10 +4,19 @@ using Autofac;
 
 namespace FacesToSmileys.Dependencies
 {
+    /// <summary>
+    /// External dependencies module
+    /// </summary>
     public class ExternalModule : Module
     {
+        /// <summary>
+        /// External dependencies
+        /// </summary>
         static IDictionary<Type, Type> Dependencies = new Dictionary<Type, Type>();
 
+        /// <summary>
+        /// Register an external dependencies
+        /// </summary>
         public static void Register<TInterface, TImplementation>() where TImplementation : TInterface
         {
             if(!Dependencies.ContainsKey(typeof(TInterface)))
@@ -16,6 +25,10 @@ namespace FacesToSmileys.Dependencies
             }
         }
 
+        /// <summary>
+        /// Load module dependencies
+        /// </summary>
+        /// <param name="builder">Builder.</param>
         protected override void Load(ContainerBuilder builder)
         {
             foreach(var dependency in Dependencies)
