--- FacesToSmileys/Models/Attitude.cs
@@ -1,15 +1,0 @@
-﻿using System;
-namespace FacesToSmileys.Models
-{
-    public enum Attitude
-    {
-        Anger,
-        Contempt,
-        Disgust,
-        Fear,
-        Happiness,
-        Neutral,
-        Sadness,
-        Surprise
-    }
-}
