--- FacesToSmileys/Models/Point.cs
@@ -1,15 +1,0 @@
-﻿using System;
-namespace FacesToSmileys.Models
-{
-    public class Point
-    {
-        public float X { get; }
-        public float Y { get; }
-
-        public Point(float x, float y)
-        {
-            X = x;
-            Y = y;
-        }
-    }
-}
