--- FacesToSmileys/Services/IFileService.cs
@@ -1,7 +1,0 @@
-﻿namespace FacesToSmileys.Services
-{
-    public interface IFileService
-    {
-        byte[] Load(string filename);
-    }
-}
