--- FacesToSmileys/Services/IPhotoService.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace FacesToSmileys.Services
-{
-    public interface IPhotoService
-    {
-        Task<byte[]> TaskPhotoAsync();
-    }
-}
