--- FacesToSmileys/Services/Implementations/FileService.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.IO;
-using System.Reflection;
-
-namespace FacesToSmileys.Services.Implementations
-{
-    public class FileService : IFileService
-    {
-        public byte[] Load(string filename)
-        {
-            byte[] bytes;
-            using (Stream s = GetType().GetTypeInfo().Assembly.GetManifestResourceStream(filename))
-            {
-                bytes = new byte[s.Length];
-                s.Read(bytes, 0, (int)s.Length);
-            }
-            return bytes;
-        }
-    }
-}
